backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
cachecontrol==0.14.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
cryptography==46.0.3
debugpy==1.8.17
decorator==5.2.1
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.1
fastapi==0.124.0
fastapi-cli==0.0.16
filelock==3.20.0
firebase-admin==7.1.0
fsspec==2025.12.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==1.2.1
hyperframe==6.1.0
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.6.3
jupyter-core==5.9.1
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
msgpack==1.1.2
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
prompt-toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
starlette==0.50.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
